+++ bssopenapi-20230930/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-03-25 Version: 1.6.3
+- Update API DescribeCoupon: add request parameters MaxResults.
+- Update API DescribeCoupon: add request parameters NextToken.
+
+
 2026-02-09 Version: 1.6.2
 - Generated python 2023-09-30 for BssOpenApi.